feat: refactor infra modules patterns to decouple the factory creation from the global files to specific libraries classes implementations
