+++ server/src/common/logger/index.ts
@@ -1,5 +1,3 @@
 export * from './logger.interface';
 export * from './logger.types';
-export * from './create-logger';
-export * from './winston-logger';
 export * from './logger';
