+++ server/src/common/logger/winston-logger.types.ts
@@ -1,0 +1,13 @@
+import { LoggerLevel } from './logger.types';
+
+export interface WinstonLoggerConfig {
+  level: LoggerLevel;
+  enableConsole: boolean;
+  serviceName: string;
+  paths: LoggerPaths;
+}
+
+interface LoggerPaths {
+  errorLog: string;
+  combinedLog: string;
+}
