+++ server/src/common/password-hasher/bcrypt-password-hasher.ts
@@ -1,11 +1,16 @@
 import bcrypt from 'bcrypt';
 import { IPasswordHasher } from './password-hasher.interface';
+import { BcryptPasswordHasherConfig } from './bcrypt-password-hasher.types';
 
 export class BcryptPasswordHasher implements IPasswordHasher {
   private readonly saltRounds: number;
 
-  constructor(saltRounds: number) {
+  private constructor(saltRounds: number) {
     this.saltRounds = saltRounds;
+  }
+
+  public static create(config: BcryptPasswordHasherConfig): BcryptPasswordHasher {
+    return new BcryptPasswordHasher(config.saltRounds);
   }
 
   hashPassword(password: string): Promise<string> {
