+++ server/src/common/password-hasher/bcrypt-password-hasher.types.ts
@@ -1,0 +1,3 @@
+export interface BcryptPasswordHasherConfig {
+  saltRounds: number;
+}
