--- server/src/common/password-hasher/create-password-hasher.ts
@@ -1,10 +1,0 @@
-import { IPasswordHasher } from './password-hasher.interface';
-import { BcryptPasswordHasher } from './bcrypt-password-hasher';
-
-interface CreatePasswordHasherConfig {
-  saltRounds: number;
-}
-
-export function createPasswordHasher(config: CreatePasswordHasherConfig): IPasswordHasher {
-  return new BcryptPasswordHasher(config.saltRounds);
-}
