+++ server/src/common/password-hasher/index.ts
@@ -1,3 +1,2 @@
 export * from './password-hasher.interface';
-export * from './create-password-hasher';
 export * from './password-hasher';
