+++ server/tests/unit/common/logger/winston-logger.test.ts
@@ -1,11 +1,12 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 import winston from 'winston';
-import { createLogger, LoggerLevel, WinstonLogger } from '@/common/logger';
+import { LoggerLevel } from '@/common/logger';
+import { WinstonLogger } from '@/common/logger/winston-logger';
 import {
   mockCommonConfig,
   mockDevelopmentConfig,
   mockProductionConfig,
-} from './create-logger.mock';
+} from './winston-logger.mock';
 
 jest.mock('winston', () => ({
   createLogger: jest.fn(),
@@ -24,13 +25,13 @@ jest.mock('winston', () => ({
 }));
 const mockWinston = jest.mocked(winston);
 
-describe('createLogger', () => {
+describe('WinstonLogger.create', () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
   it('should create a logger for every environment', () => {
-    const logger = createLogger(mockCommonConfig);
+    const logger = WinstonLogger.create(mockCommonConfig);
 
     expect(logger).toBeInstanceOf(WinstonLogger);
     expect(logger.info).toBeDefined();
@@ -41,13 +42,13 @@ describe('createLogger', () => {
   });
 
   it('should call winston.createLogger', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.createLogger).toHaveBeenCalled();
   });
 
   it('should call winston.createLogger with the correct config for every environment', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.createLogger).toHaveBeenCalledWith(
       expect.objectContaining({
@@ -58,7 +59,7 @@ describe('createLogger', () => {
   });
 
   it('should create a logger with the correct format for every environment', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.format.combine).toHaveBeenCalled();
     expect(mockWinston.format.timestamp).toHaveBeenCalledTimes(1);
@@ -68,7 +69,7 @@ describe('createLogger', () => {
   });
 
   it('should use the correct error log path for every environment', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.transports.File).toHaveBeenCalledWith({
       filename: mockCommonConfig.paths.errorLog,
@@ -77,7 +78,7 @@ describe('createLogger', () => {
   });
 
   it('should use the correct combined log path for every environment', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.transports.File).toHaveBeenCalledWith({
       filename: mockCommonConfig.paths.combinedLog,
@@ -85,25 +86,25 @@ describe('createLogger', () => {
   });
 
   it('should use two file transports for every environment', () => {
-    createLogger(mockCommonConfig);
+    WinstonLogger.create(mockCommonConfig);
 
     expect(mockWinston.transports.File).toHaveBeenCalledTimes(2);
   });
 
   it('should use the console transport when enableConsole is true in development', () => {
-    createLogger(mockDevelopmentConfig);
+    WinstonLogger.create(mockDevelopmentConfig);
 
     expect(mockWinston.transports.Console).toHaveBeenCalled();
   });
 
   it('should not use the console transport when enableConsole is false in production', () => {
-    createLogger(mockProductionConfig);
+    WinstonLogger.create(mockProductionConfig);
 
     expect(mockWinston.transports.Console).not.toHaveBeenCalled();
   });
 
   it('should use the correct format for the console transport when enableConsole is true in development', () => {
-    createLogger(mockDevelopmentConfig);
+    WinstonLogger.create(mockDevelopmentConfig);
 
     expect(mockWinston.format.combine).toHaveBeenCalledTimes(2);
     expect(mockWinston.format.colorize).toHaveBeenCalled();
@@ -111,7 +112,7 @@ describe('createLogger', () => {
   });
 
   it('should not call the extra format functions when enableConsole is false in production', () => {
-    createLogger(mockProductionConfig);
+    WinstonLogger.create(mockProductionConfig);
 
     expect(mockWinston.format.combine).toHaveBeenCalledTimes(1);
     expect(mockWinston.format.colorize).not.toHaveBeenCalled();
