--- server/tests/unit/common/password-hasher/create-password-hasher.test.ts
@@ -1,15 +1,0 @@
-import { describe, expect, it } from '@jest/globals';
-import { BcryptPasswordHasher } from '@/common/password-hasher/bcrypt-password-hasher';
-import { createPasswordHasher } from '@/common/password-hasher/create-password-hasher';
-
-const config = {
-  saltRounds: 10,
-};
-
-describe('createPasswordHasher', () => {
-  it('should create a bcrypt password hasher', () => {
-    const passwordHasher = createPasswordHasher(config);
-
-    expect(passwordHasher).toBeInstanceOf(BcryptPasswordHasher);
-  });
-});
